% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_intersimR}
\alias{hello_intersimR}
\title{Hello from intersimR}
\usage{
hello_intersimR()
}
\value{
A character string.
}
\description{
Minimal function used to make the package pass R CMD check
while the core API is developed.
}
\examples{
hello_intersimR()
}
