% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersimR-package.R
\docType{package}
\name{intersimR-package}
\alias{intersimR}
\alias{intersimR-package}
\title{intersimR: Simulation-based detection of animal–vessel interaction events}
\description{
Tools to detect and quantify local interaction events from animal and vessel
tracking data using simulation-based null models (attraction, following, etc.).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spatialmarine.github.io/intersimR/}
}

}
\author{
\strong{Maintainer}: David March \email{david.march@uv.es} (\href{https://orcid.org/0000-0002-6118-761X}{ORCID})

}
\keyword{internal}
