% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-class.R
\name{is_track_tbl}
\alias{is_track_tbl}
\title{Test if an object is a track table}
\usage{
is_track_tbl(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Test if an object is a track table
}
