% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-point-check.R
\name{make_point_check}
\alias{make_point_check}
\title{Create a point.check function for availability::surrogateAR()}
\usage{
make_point_check(oceanmask)
}
\arguments{
\item{oceanmask}{terra::SpatRaster with values 0=ocean, 1=land}
}
\value{
function(tm, pt) -> logical (TRUE if at sea)
}
\description{
Create a point.check function for availability::surrogateAR()
}
