% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-animal-vessels.R
\name{match_animal_vessels}
\alias{match_animal_vessels}
\title{Match animal fixes to vessel positions within time and distance thresholds}
\usage{
match_animal_vessels(animal, vessel, dist_thr_m, time_thr_min)
}
\arguments{
\item{animal}{track_tbl with columns: id, lon, lat, time, fixID}

\item{vessel}{track_tbl with columns: id, lon, lat, time}

\item{dist_thr_m}{numeric. Distance threshold in meters}

\item{time_thr_min}{numeric. Half time window in minutes (±)}
}
\value{
A tibble with columns:
matchID, fixID, animalID, vesselID,
animal_time, vessel_time,
animal_lon, animal_lat, vessel_lon, vessel_lat,
dt_min, dist_m
Attributes:
dist_thr_m, time_thr_min
}
\description{
Match animal fixes to vessel positions within time and distance thresholds
}
