% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-class.R
\name{print.track_tbl}
\alias{print.track_tbl}
\title{Print method for track tables}
\usage{
\method{print}{track_tbl}(x, ...)
}
\arguments{
\item{x}{A \code{track_tbl}.}

\item{...}{Ignored.}
}
\description{
Print method for track tables
}
