% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegap-segments.R
\name{timegap_segments}
\alias{timegap_segments}
\title{Segment a sorted time series by a maximum gap (in minutes)}
\usage{
timegap_segments(time, max_gap_min)
}
\arguments{
\item{time}{POSIXct vector, already sorted}

\item{max_gap_min}{numeric, gap threshold in minutes}
}
\value{
integer segment index starting at 1
}
\description{
Segment a sorted time series by a maximum gap (in minutes)
}
